[refactor] #52 : service hooks 분리
